404 ERROR for delete, post, get

+++ src/main/java/com/devsuperior/Challenge03/services/ClientService.java
@@ -3,14 +3,20 @@ package com.devsuperior.challenge03.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.challenge03.dto.ClientDTO;
 import com.devsuperior.challenge03.entities.Client;
 import com.devsuperior.challenge03.repositories.ClientRepository;
+import com.devsuperior.challenge03.services.exceptions.DataBaseException;
+import com.devsuperior.challenge03.services.exceptions.SearchNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -20,14 +26,15 @@ public class ClientService {
 	
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		Client client = repository.findById(id).get();
+		Client client = repository.findById(id).orElseThrow(
+				() -> new SearchNotFoundException("Resource not found!"));
 		return new ClientDTO(client);
 	}
 	
 	@Transactional(readOnly = true)
-	public List<ClientDTO> findAll(Pageable pageable) {
+	public Page<ClientDTO> findAll(Pageable pageable) {
 		Page<Client> search = repository.findAll(pageable);
-		return search.stream().map(x -> new ClientDTO(x)).toList();
+		return search.map(x -> new ClientDTO(x));
 	}
 	
 	@Transactional
@@ -40,15 +47,28 @@ public class ClientService {
 
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
+		try {
 		Client client = repository.getReferenceById(id);
 		dtoToEntity(dto, client);
 		client = repository.save(client);
 		return new ClientDTO(client);
+		}
+		catch (EntityNotFoundException e) {
+			throw new SearchNotFoundException("Resource not found!");
+		}
 	}
 		
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new SearchNotFoundException("Resource not found");
+		}
+		try {
 		repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException("Integrity violation reference fail");
+		}
 	}
 	
 	private void dtoToEntity(ClientDTO dto, Client client) {
